net: usb: lan78xx: Enforce a minimum interrupt polling period
If a new reset event appears before the previous one has been
processed, the device can get stuck into a reset loop. This happens
rarely, but blocks the device when it does, and floods the log with
messages like the following:

  lan78xx 2-3:1.0 enp1s0u3: kevent 4 may have been dropped

The only bit that the driver pays attention to is "link was reset". If
there's a flapping status bit in that endpoint data (such as if PHY
negotiation needs a few tries to get a stable link), polling at a
slower rate allows the state to settle.

See also: https://github.com/raspberrypi/linux/issues/2447

Signed-off-by: Fiona Klute <fiona.klute@gmx.de>
Cc: kernel-list@raspberrypi.com
Cc: stable@vger.kernel.org